Update to Windows Form
Windows Forms give the developer the possibility to implement interface
functions. Here for example is minimize to notify implemented.
