+++ mi/mi.sln
@@ -1,0 +1,22 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 14
+VisualStudioVersion = 14.0.25123.0
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "mi", "mi.csproj", "{3D80D830-B531-44CC-A74E-2B2D1C2F0F16}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{3D80D830-B531-44CC-A74E-2B2D1C2F0F16}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3D80D830-B531-44CC-A74E-2B2D1C2F0F16}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3D80D830-B531-44CC-A74E-2B2D1C2F0F16}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3D80D830-B531-44CC-A74E-2B2D1C2F0F16}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
